tests/wpt/tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.9arg.destsize.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.drawImage.9arg.destsize</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.9arg.destsize</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  const response_red = await fetch('/images/red.png');
  const blob_red = await response_red.blob();
  const bitmap_red = await createImageBitmap(blob_red);

  const response_green = await fetch('/images/green.png');
  const blob_green = await response_green.blob();
  const bitmap_green = await createImageBitmap(blob_green);

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.drawImage(bitmap_green, 1, 1, 1, 1, 0, 0, 100, 50);
  ctx.drawImage(bitmap_red, 0, 0, 100, 50, -50, 0, 50, 50);
  ctx.drawImage(bitmap_red, 0, 0, 100, 50, 100, 0, 50, 50);
  ctx.drawImage(bitmap_red, 0, 0, 100, 50, 0, -25, 100, 25);
  ctx.drawImage(bitmap_red, 0, 0, 100, 50, 0, 50, 100, 25);
  _assertPixelApprox(canvas, 0,0, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 99,0, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 0,49, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 99,49, 0,255,0,255, 2);

}, "");
</script>
